extract actual file IO functions
